create main pkg
