+++ app/__init__.py
@@ -2,13 +2,9 @@ from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate
 
-# from config import Config
-
 
 db = SQLAlchemy()
 migrate = Migrate()
-
-# from app import routes
 
 
 def create_app():
@@ -19,9 +15,10 @@ def create_app():
     db.init_app(app)
     migrate.init_app(app, db)
 
-    # from app import routes
     from app.api import bp as api_bp
+    from app.main import bp as main_bp
 
     app.register_blueprint(api_bp, url_prefix="/api")
+    app.register_blueprint(main_bp, url_prefix="/")
 
     return app
